move "find best room" algo to slab_room_detection.c
This makes it easier to replace/improve later.
Any functions not moved to slab_room_detection should be suitable for any implmentation.
