+++ Makefile
@@ -264,6 +264,7 @@ obj/room_util.o \
 obj/room_workshop.o \
 obj/scrcapt.o \
 obj/slab_data.o \
+obj/slab_room_detection.o \
 obj/sounds.o \
 obj/spdigger_stack.o \
 obj/tasks_list.o \
